requirements.txt
google-cloud-vision==3.1.4
google-cloud-core==2.3.2
google-cloud-storage
google-cloud-dlp
opencv-python
pillow
Flask==2.2.2
flask-cors
gunicorn==20.1.0
onerecord==0.2.0
opencv-python